Rename simple sketch with cache
